Create location domain and data modules with the support of CRUD operations

+++ domain/src/main/java/cs/vsu/ru/domain/usecase/location/RemoveSavedLocationUseCase.kt
@@ -5,6 +5,15 @@ import cs.vsu.ru.domain.repository.location.LocationRepository
 class RemoveSavedLocationUseCase(private val locationRepository: LocationRepository) {
 
     fun execute(locationName: String) {
+        // TODO: if removing current, favorite or last saved location add default location
+
+        if (locationRepository.getSavedLocations().size == 1 ||
+            locationRepository.getFavoriteLocation()?.name == locationName ||
+            locationRepository.getCurrentLocation()?.name == locationName
+        ) {
+            throw Exception("gfregerhger")
+        }
+
         locationRepository.removeSavedLocation(locationName)
     }
 }
